feat: strict_type=1

+++ modules/demo/src/App/Forms/FormGridPoppyEstablish.php
@@ -25,7 +25,7 @@ class FormGridPoppyEstablish extends FormWidget
 
     public function handle()
     {
-        $id   = input('id');
+        $id   = (int) input('id');
         $Area = new Area();
         if (is_post()) {
             if ($Area->establish(input(), $id)) {
